+++ CHANGES.md
@@ -15,6 +15,7 @@ Supported Metrics:
 * Fixed #348: String index out of range: -2. (@timrsfo & Nick Tan)
 * Fixed #407: eclipse plugin generates invalid java for toString & hashCode (@dabraham02124 & Nick Tan)
 * Fixed #408: Doesn't build against or run on Eclipse 4.10 (2018-12)
+* Fixed #410: After adding module-info.java, testng-eclipse no longer finds test classes in maven project. Aka, support Java 11.
 
 ## 6.14